move: 공용 inputs 컴포넌트 위치 Auth 폴더 하위로 변경
