+++ src/components/common/Auth/Inputs/InputPassword/index.tsx
